Convert to genderize.io API (and using gendered gem)
Now using a much better API that is based on worldwide data, and is thus doing much better with foreign names (previous API was based in 2010 U.S. census data). Have written the algorithm to batch 10 names at a time which only uses 1 request as opposed to 10. I get 1000 requests a day, so an upper bound of 10,000 RSVPs.

Still, I wonder if the right move is actually to save meetup users in a database. That way we don't have to look them up if we've asked Genderize for the prediction previously, and it would support features later like crowd-sourcing gender corrections ("That's not a man!")

Also created Ruby RSVP objects, which I should have done from the beginning.

Happy with it so far. Next steps could be: refactoring, writing tests, validating user input, preparing for errors/bad outcomes like running out of genderize requests, visualizations, and supporting analysis for an organization instead of just an event.