Add basic d3 chart; not sure if successful
So I ran out of requests on genderize.io already... I hope that's just because of my inefficient testing yesterday. So I was not able to verify that the chart actually appears as desired and am just calling it for today until I get some more requests.

Long-term strategy: save results of the API call to the database, so that you can only run the API call for names you've never predicted before.