Get basic percent chart working
Basic d3 now up and running. Would be nice to make it prettier.

I also discovered that the call to API is still charging one request per name (I thought we were doing multiple names per request). This is making me realize that the only way this app is viable, even in the short-term, is to save the results of the API call to a database so we only make requests for names we've never seen before. I'll try to do this tomorrow.