Convert to pgsql, add names table, start saving API calls
We now save the results of API calls to a database so that we only have to spend an API request once - the first time we come across that name. This will make the app much more likely to survive moderate levels of traffic, as presumably by the time we get there, the most common names will be in the database.

Heroku's table limit of 10000 names may eventually be a pain, but we'll cross that bridge when we come to it (my first bid would be to only save common names and delete uncommon names to make room).

I'm going to deploy to Heroku, but then it's definitely time to build out the test suite and refactor - have procrastinated on that too long.